Add form to delete an activity
Previously it wasn't possible to delete your own activities, which bit
me once when I created a test activity by accident on the production
system instead of locally.

This adds a danger zone area at the bottom of the edit form, with a
checkbox that needs to be checked to confirm the deletion.